+++ src/rust/luisa_compute_cpu_kernel_defs/cpu_kernel_defs.h
@@ -55,9 +55,9 @@ using OnHitCallback = void(*)(RayQuery*);
 
 struct Accel {
     const void *handle;
-    Hit (*trace_closest)(const void*, const Ray*, uint8_t);
-    bool (*trace_any)(const void*, const Ray*, uint8_t);
-    void (*set_instance_visibility)(const void*, uint32_t, uint8_t);
+    Hit (*trace_closest)(const void*, const Ray*, uint32_t);
+    bool (*trace_any)(const void*, const Ray*, uint32_t);
+    void (*set_instance_visibility)(const void*, uint32_t, uint32_t);
     void (*set_instance_transform)(const void*, uint32_t, const Mat4*);
     void (*set_instance_user_id)(const void*, uint32_t, uint32_t);
     Mat4 (*instance_transform)(const void*, uint32_t);
